+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Added Form\UploadProfilePictureForm test
 - Added Options\ModuleOptions test
 - Added Module test
+- Added Service\ProfilePicture test